Implement JWT Authentication

+++ src/main/java/com/bvr/LearningProject/repository/AccountsRepository.java
@@ -4,10 +4,9 @@ import com.bvr.LearningProject.model.Accounts;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-
 @Repository
 public interface AccountsRepository extends CrudRepository<Accounts, Long> {
-	
-	Accounts findByCustomerId(int customerId);
+
+    Accounts findByCustomerId(long customerId);
 
 }
