+++ src/main/java/com/bvr/LearningProject/repository/CardsRepository.java
@@ -2,6 +2,7 @@ package com.bvr.LearningProject.repository;
 
 import java.util.List;
 
+
 import com.bvr.LearningProject.model.Cards;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
@@ -10,6 +11,6 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface CardsRepository extends CrudRepository<Cards, Long> {
 	
-	List<Cards> findByCustomerId(int customerId);
+	List<Cards> findByCustomerId(long customerId);
 
 }
