+++ src/main/java/com/bvr/LearningProject/repository/LoanRepository.java
@@ -10,6 +10,6 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface LoanRepository extends CrudRepository<Loans, Long> {
 	
-	List<Loans> findByCustomerIdOrderByStartDtDesc(int customerId);
+	List<Loans> findByCustomerIdOrderByStartDtDesc(long customerId);
 
 }
